In response to bug 2073 Added instruction to list: "Is your data available on GEO? If so, it is probably already loaded for you. Check here." With a link to the dataset listing.
Also, added a pop-up message that appears when a user is loading in a dataset and uses something similar to a GEO ID for 'name' or 'short name'. The message asks them to check if their datase is already available in Gemma. Similarity is defined as the letters 'gse' followed by 3 to 6 digits.