(Bug 3221 Added a method that returns the external database statistics)
We need a page that lists all of Neurocarta external databases as well as some
summary statistics (# of evidence, genes, phenotypes) and last time it was
updated.

It could be a new link in the about section "Neurocarta external sources
updates".

It is fairly urgent because we need it up before we send the manuscript back to
the reviewers. I'm not sure whether it would be better for Nicolas or Frances
to do it. You can decide.