Fixes missing taglibs
The taglibs weren't necessarily missing as this jsp is intended to be
combined  into another file with the required taglibs. This, however,
means that eclipse will complain (rightly so) about them as nothing is
stopping anyone from using this file without the required taglibs.

The taglib redundancy is a non-issue.