Moves shared dependency baseCode to parent pom
The reasoning behind this move is that maven's dependency conflict
resolution uses a nearest-wins strategy. For some of the gemma
sub-projects the nearest-winning baseCode dependency resided in
something like GEOMMTx (example gemma-web). This causes weird issues
when attempting to update baseCode versions across the entire Gemma
project as some would use the baseCode in gemma-model while others the
baseCode in GEOMMTx. This solves that issue by having the baseCode in
the parent pom always be the nearest.