Fixes Bug 4420 - NCBI switch to https
Lots of xml loaders were using the http protocol uris. Even though the
http connections were being correctly redirected to https on NCBI's end,
URLConnection does not follow redirects that cross protocols (by
design). 

See
http://docs.oracle.com/javase/6/docs/technotes/guides/deployment/deployment-guide/upgrade-guide/article-17.html

Note that there are a few changes to XML parsers as well. This has to do
with the use of inputStream.available() to check for empty xml streams.
This is not the intended use-case for available().
inputStream.available() tells you if there is something ready to be read
without blocking. And it always returns zero in some cases, e.g. SSL
sockets. It doesn't tell you how much data is in the input stream. This
caused failing tests as the new https enabled connections blocked on
first read.