Fixes unnecessary hibernate updates on session flush
Hibernate was marking entities with null database values on any property
of UserType
ubic.gemma.persistence.hibernate.usertypes.HibernateStringClobType as
dirty regardless of if said entity changed in the session or not.

This appears to be caused by a combination of incorrect hashcode, equals
and nullsafeget/set implementations. Hibernate would see the previous
state as an empty string and the current state as null, thus flagging
the entity as dirty.

Instead of fixing the custom UserType, I opted to use the inbuilt
Spring-Hibernate solution of
org.springframework.orm.hibernate3.support.ClobStringType which mirrors
our use-case.

This calls into question our custom UserType HibernateByteBlobType. It
can potentially be replaced by Spring-Hibernate BlobByteArrayType, but I
will leave this for further investigation.