Fixes 4440 - missing mean-variance plots. Also plotting all data now.
The problem was that lowess vectors were shorter than mean & variance
vectors, which caused problem in a for loop in an "outlier"-removal
method. I added conditions to the for loop, but we are also going to
look into why the vector is actually shorter. Also we are now going to
try to displaying all the data, so the remove-outliers method is not
going to be used at all.