cli: Improve error handling and conversion to exit codes (fix #79)
Make doWork throw an exception instead of returning it. This
substantially simplify the implementations since we don't have to check
the return value.

Move executeCommand into AbstractCLI and use it systematically to launch
the CLI.

Make executeCommand the ultimate handler of exceptions to raise failure
exit code (1).

Make executeCommand check for error objects and raise a different
failure exit code (2).