cli: Make successObjects and errorObjects private
Add addSuccessObject and addErrorObject methods additional information:

 - an batch-processed object
 - a message
 - a Throwable to produce a meaningful stacktrace

Remove usage of log.error or log.info when the message is already
appended to the error or success objects since it will be displayed when
summarizing the results.

Always call summarizeProcessing to simplify doWork even if a failure
occurs. If no object has been processed, nothing will be printed anyway.