cli: Use an ExecutorService for processing parallel batch tasks
Use a centralized ExecutorService to process batch tasks to simplify how
individual commands handle parallelism.

Add a minimal executeBatchTasks() routine in AbstractCLI to process
tasks and retrieve their results.

Use a ForkJoinPool which is more efficient than running a bunch of
consumer threads with a single blocking queue. We can get a similar
pattern with ThreadPoolExecutor with the advantage of reusing an
optimized thread pool implementation.