Fix test ehcache configuration for ExpressionAnalysisResultSet
This is a continuation of a76fe0829da64fbb78ef2f40ba4843ad49dcbc13.

The reason this failed is that there used to be two cache entries for
ExpressionAnalysisResultSet, and I accidently refactored back and forth
the class from the diff subpackage and it resulted in the cache entry
being renamed.

The fix is to remove the redundant entry, which should have never been
there in the first place.