Fix offset/limit in datasets and platforms endpoints (fix #164) [WIP]
Break-up VO filtering in a separate interface since only a few of our
DAO and services actually implement it.

Use entity-driven HQL queries so that the offset/limit applies to the
entities instead of the raw SQL result set.

Add a Slice<T> structure to represent an offset/limit of a larger
List<T> like most Hibernate results to eventually provide more
information to the API.