Fix comparison by string in filtering (fix #181, #175)
An incorrect value was returned in the case of a string.

Make the problematic routines in ObjectFilter static, they should never
had access to the requiredValue in the first place.

Adjust the generated SQL to avoid parenthesizing non-collection
bindings so they won't be mistreated as sets.

Add a test case with with filtering EE by short name.