Improve API ergonomics
Add a processLoadValueObjectsQueryResult protected routine to
AbstractFilteringVoEnabledDao so that more of the querying logic can be
shared.

Add FilteringDao and FilteringService interfaces for DAO that do not
fully implement VO filtering, but still want to provide some filtering
capabilities on their own routines.

Add a prototype for the SQL generator based on Hibernate Criteria API
that is currently being experiment with for the result sets endpoint.

Introduce OffsetArg and LimitArg to parse offset and limit and produce
errors if invalid values are provided.

Add tests for FilterArg and ObjectFilter.