rest: Fix filtering analysis result sets
Finish the implementation of filtering analysis result sets:

 - use null as object alias to refer to the root entity
 - use long for public SID
 - use a detached query, select the AclObjectIdentity that match the
   entity being requested and test it with Subqueries.propertyIn

Finish the implementation of ACL on the associated EE.

Fix tests that used to pass null to FilterArg where an empty filter
would be the appropriate value.

Add extra tests for various aspects of the filtering such as testing if
an id is within a given set.