Fix fetching taxa by scientific name (fix #195)
Rework getObjectFilter to return a disjunction clause instead of a
single filter because the taxon rely on testing both scientific name and
common name.

In the TaxonArg implementation, we simply override the default behaviour
to construct a clause containing both the common name and the scientific
name.

Remove combineFilters in favour of building the filter array in the
controller.