Make thawing part of BaseDao and BaseService and introduce a more efficient loadAndThaw
The main problem with thawing is that it is often performed in a context
where the original Hibernate session that was used to load the model is
already closed, thus forcing a reload of the entity in order to thaw it.

Instead, propose a new interface that performs loading and thawing in a
single database transaction and deprecated thaw() pretty much
everywhere.

We also replace the various existing implementation of thaw() by a
single implementation in AbstractService. This new implementation thaws
a reloaded entity (or entities for the collection case) and return it.

Note that thawing in the DAO layer is not deprecated at all, and
actually highly desirable to have it there. Unlike in the service layer,
it's perfectly fine to assume that thaw() in the DAO is performed in the
same session as any previous load().

The counterpart of thaw() in the DAO operates in-place with the same
spirit that Hibernate.initialize() involves. Thawing collection can be
made more efficient by requerying, and this is not prevented as a
specific implementation can clear() and addAll() the input collection.

Use Set<T> in model definitions when unicity is expected. This is
basically triggered by the necessity to specialize some Collection<T> to
List<T> so that they can be thawed in the DAO layer.