Reuse the same VO conversion logic for the Web interface and RESTful API
Move GeneralSearchController's VO conversion logic into SearchService so
that it can be reused for the RESTful API.

Add a new ValueObjectConverter that implements Spring Converter
interface to apply VO conversion logic defined in BaseVoEnabledService.
This is used to rewrite the logic of converting VOs.

Add tests for GenericValueObjectConverter

Make conversion more flexible for source/target types