Rewrite the task progress update callback API
Invert the logic of progress update by allowing a Log4j appender to
invoke a configured callback instead of attaching a new appender for
every task context.

This approach requires that we pre-configure the appender in log4j.properties,
which we do for tests and gemma-web new suggested defaults.

Pass -Dlog4j1.compatibility=true to surefire so that the two tests that
now require it run properly.