pom: Move schema export task under its own execution
As defined, the target is inherited and then shadowed in the
version-file execution.

Add schema-export as execution <id/>.

Remove ${maven.dependency.classpath} mention as it was removed in
maven-antrun-plugin 3.0.0. Instead, add hibernate-core to the plugin
dependencies. This limit the scope of the classpath the plugin has
access to.

Add ${project.basedir} prefix to the *.hbm.xml file path resolution.