test: Simplify BaseSpringContextTest
Remove getBean() in favour of @Autowired. It's entirely equivalent, so
there is no need for the helper.

Remove sqlScriptEncoding and use Charsets.UTF_8 instead.

Remove setDataSource() and perform intialization of jdbcTemplate in
afterPropertiesSet().

Perform authentication as admin in setUp() so that all tests run with a
deterministic user.

Fix tests using setup() instead of setUp() and call super.setUp() ensure
that the current user is granted admin privilege.