Remove HibernateDaoSupport and HibernateTemplate usages
Usage of HibernateDaoSupport is currently the only thing blocking us
from moving to Hibernate 4 while remaining with Spring 3 as this
configuration only support the SessionFactory injection.

Inject a SessionFactory wherever a HibernateDaoSupport is used and
replace calls with equivalent calls.

Replace a few list().iterator().next() approaches with uniqueResult().

Replace doInHibernate() with a @Transactional-annotated method.

Replace convertHibernateAccessException() usages to direct call as the
majority of code does.