Fix generic type usages for SearchResult and nullable for the search package
Introduce a SearchResultMap interface that provides a typed get() so
that no casting is necessary when retrieving results by type.

Remove empty list initialization in favour of defaulting to an empty
list. Rely on SearchSettings to determine if a certain result type is
desirable instead of testing if its results collection has been
initialized.

Prefer simply SearchResult<?> over SearchResult<? extends Identifiable>
since the latter is implied in the type definition.