Use persist() instead of save()
Unlike save(), persist() is part Java Persistence API.

It also does not guarantee to generate an ID, although in practice we
always have one since our primary keys use an auto-increment strategy.

Remove the return values of BaseDao.create() and BaseService.create() as
creation is now an in-place operation that attaches the instance to the
Hibernate session.