Improve homologene service initialization
Use an bean factory pattern to initialize HomologeneService on-demand.

Introduce a new AsyncFactoryBean interface and base implementation to
create bean asychronously. This implementation works for singleton and
non-singleton beans and uses Future.cancel() on destroy instead of using
a daemon thread. This allows for clean shutdown behavior.

Use getObjectAsync() from AsyncFactoryBean in StartupListener to prevent
blocking while initializing the service.