Improve definition of AsyncBeanFactory
Instead of having a separate getBeanAsync() logic path, define an
AsyncBeanFactory<T> to be a BeanFactory<Future<T>>. This implies that
the getBean() operation is fundamentally asychronous and coordinated by
the factory.

Inject a Future<HomologeneService> where needed, which triggers its
asynchronous initialization. Using this, we do not need to trigger it in
StartupListener.