Add an endpoint to retrieve annotations usage statistics by dataset filter
Annotations are obtained from tags, biomaterials and factor values as
usual. The three queries are performed concurrently. If biomaterials
don't complete by two seconds, it is left to run in the background and
partial statistics are returned.

The statistics obtained are respecting ACLs. Caching works partly under
the assumption that most requests will be issued by anonymous users.

Add missing parts of the FilteringVoEnabledDao to filter at
entity-level, allowing us to ignore the VO conversion for datasets when
retrieving their annotations.