Fix resolution of filterable property meta for native enums
Now that enums are natively mapped by Hibernate, the bound parameter
must be an instance of the enumerated Java type.

Improve the way subclasses of AbstractFilteringVoEnabledDao generate
filterable property meta-information. The new approach ensures that the
resolved property retains description and available values.

Use a cache for memoizing resolved property meta and avoid costly
tree-traversal of the object graph structure.