Few improvements for computing properties meta
Use a shared static cache for filterable prop meta so that all
implementations can reuse it.

Don't render available properties anymore when an exception is raised.
It's not really useful given that those are available from the OpenAPI
specification.

Make it an error to call addFilterableProperties() with a non-strictly
positive maximum depth.