Few cleanups for AclQueryUtils
Use sid instead of sid.id, we don't need the assumption that the
identifier column is named 'id'.

Remove dependency on FilterQueryUtils in AclQueryUtils, the caller must
now provide the ID column.

Use a subquery to retrieve the anonymous sid instead of assuming that 4
is anonymous.

Fix incorrect subqueries in AclCriteriaUtils.

Add more thorough tests for all mode: admin, user and anonymous.