Keep track of the original property in Filter and Sort
Add an originalProperty field to Sort and Filter to keep track of the
property name in the original space. For example, when a user specify a
filter, it is translated to a specific object alias and property name
applicable to the HQL or criteria query. By keeping the property before
the translation, we can restitute at a later time.

Add a PropertyMapping interface to extract the shared components of
mapping some original property to an object alias and property name
pair.