Cleanups for LinearModelAnalyzer
Indicate that the overall P-value is null/NaN in the warning message.

Improve the logic for resolving the Future using Future.get() with a
timeout instead of busy waiting with Thread.sleep().

Use a single thread executor as there's no need to occupy a thread from
the local task executor since we don't return a future.