search: Retrieve EEs by inferred term in bulk
Collect all children before calling findExperimentByUris() which is
optimized for gathering EEs in bulk.

Extract ontology term children caching logic in a reusable component and
provide a fast routine for collecting all the children of a given set of
terms.

Use a proper VO for mapping search results to the frontend.