Add test for cache invalidation when a characteristic is removed
The test covers the behaviour when a term is deleted which should in
turn evict the cached query that computes annotation usage frequency.

Fix definitions of foreign keys for the EE2C table. It appears that they
have to be specified separately in an alter table statement to be
effective.