Ensure that data appearing log-transformed is inferred as LOGBASEUNKNOWN
The strategy used by findScale() now return LOGBASEUNKNOWN when the data
appears to be log-transformed. This in turn produces an exception in
ensureLog2Scale() instead of log-transforming the data twice.

Rename UnknownLogScaleException to UnsupportedScaleConversionException
to emphasis the problem that the scale cannot be converted to log2.

Add some tests to validate various conversion strategy.