Fix duplicated GEEQ entity when updating scores
Change the signature of calculateScore() to take an entity as input
instead of an EE ID and an enum for scoring modes.

The implementation of GeeqService does not need to thaw anything since
it can operate within a single Hibernate session.

Fix unintended rollback when computing GEEQ score of a dataset that has
missing processed EVs. Eliminate the NoProcessedExpressionDataVectorsException
in favour of simply returning null, which triggers no rollback.