Fix removal and thawing of EEs
Favour in-query fetching via join over running multiple selects for
one-to-one relations.

In remove, use merge() to make sure the EE is properly attached to the
session.

Add cascade="delete,delete-orphans" to EE-associated QTs and remove the
bits from remove().

Fix EE split tests: parts are removed when the EE is removed, so there's
no need to remove parts individually. Also address cleanups of already
loaded EEs.

Remove DataUpdater.deleteData() which is redundant with removing an EE
and results in QTs being deleted twice in the test case.