Ensure that audit events with exceptions are recorded
Audit events are usually not recorded when the transaction is rolled
back. This is problematic for the pattern of creating an audit event
from an exception and then raising the said exception which would result
in a rollback.

To prevent this, annotate addUpdateEvent with REQUIRES_NEW propagation
so that an entirely new transaction is used for the purpose of creating
the audit event.

Consistently use AuditTrailService to create audit events.

Cleanup unused methods in AuditTrail.