Prepare for gsec 0.0.12
Remove ACL checks where the return type or input type is an ID since
this is not handled and will result in an error. A Securable is now
always expected.

Move ACLs-related annotations at the service layer.

Don't use `Collection<? extends ...>` when filtering with ACLs since its
element is not assignable to a Securable.

Never allow users to remove entities by ID since no ACL check can be
performed.