Add support for multiple highlights
Remove doHighlighting and simply check if a highlighter is non-null in
the search settings.

Use a mapping for representing highlights so that any matched field in
the entity can be discovered in the results.

Trim Compass highlights which sometimes contain leading spaces that do
not appear in the original text.

Use a break to separate highlights for multiple terms