Type-safe parsing of GEO records
Use a BiConsumer<U, T> and refer to the actual setters when parsing GEO
records. This is safer as we don't need to be worried with missing
fields.

Add missing getters and setters in the data model.

Fix a bug when setting the fax number of a sample in a series and
platform.