Fix characteristic browser (fix #773)
Retrieve the characteristics from the CHARACTERISTIC table in a single
query, taking advantage of the foreign key structure to subsequently
retrieve their parents.

Limit the number of results in the query itself and add capabilities for
restricting results to a specific set of parent classes.

Fix unintended proxy initialization when generating links.

Don't navigate parents, it's too expensive.