rest: Add private to Cache-Control header
Make maxAge optional and default to -1 (i.e. does nothing).

Add two attributes to the annotation: isPrivate and authorities.

Combined, these two features allow one to express that an endpoint can
be cached, but if the content is personalized, then make it private.

For example:

    @CacheControl(maxAge = 1200)
    @CacheControl(isPrivate = true, authorities = {"GROUP_USER"})
    public String endpoint();

Is valid for 1200 seconds and if the current user is logged in, then it
is also private and cannot be cached by intermediate servers.