Introduce a Statement model for FactorValue characteristics
This is based of the discussion I had with Paul to use a simple approach
of subclassing Characteristic so that it can hold up to two related
terms that may or may not be part of the bag of terms.

The CHARACTERISTIC table has 7 added columns: class, PREDICATE,
PREDICATE_URI, OBJECT_FK, SECOND_PREDICATE_URI and SECOND_OBJECT_FK.

The class defaults to NULL for regular characteristics and 'Statement'
for statement-like characteristics. For the moment, it is only being
used by FactorValue, but other entities might benefit from added
semantic in the future.