Use better exception in Gemma Web
Use EntityNotFoundException and IllegalArgumentException to produce 404
and 400 errors.

Consistently use EntityNotFoundException when an entity is missing.

Fix some of the controller signature so that we don't need to rely on
HttpServletRequest but instead use Spring to extract request parameters.

Add two loadOrFail() variant in BaseService to allow passing a custom
exception and a custom exception with a message.