Eliminate offset/limit parameters and improve the code for retrieving analysis VOs
Add a parameter to include analyses of subsets.

Include analyses of subsets in the output of the endpoint.

Add two new service methods to retrieve analysis VOs: one by a
collection of BioAssaySet and another one for a single BioAssaySet. Use
the latter for the REST API.

Rename sourceExperiment to sourceExperimentId and retain the older name
marked as deprecated (fix #844).