Adapt warning depending on the profile
Warn on the CLI if ontologies or Homologene is enabled on startup and do
the exact opposite for Gemma Web.

Only warn that load.ontologies is false if at least one ontology is
enabled.

Warn on the CLI if Hibernate is configured to validate the database on
startup. This is unnecessary and costly.

Remove redundant GO service warning if load.ontologies is false, this is
already covered by the ontology service.

Make the GEO browser bean lazy. It's only used in Gemma Web, so it won't
be loaded on the CLI.