Update baseCode to 1.1.19 (fix #842)
Update baseCode to 1.1.19 which adds support for setting language levels
and inference mode.

Allow language level and inference mode to be fine-tuned in
OntologyServiceFactory. The combination of the two has a significant
incidence on loading time, so one must carefully set them.

Add support for setting additional property URIs to be used for
inferring parents and children.

Add tests to ensure that "substantia nigra pars compacta" is inferred
from "brain".