Fix substitution of properties across different sources.
Remove unused configuration entries in default.properties.

Use /var/tmp/gemmaData as a more sensible default.

Populate the environment with the settings so that they also become
available via @Value("#{environment['property']}") and Environment.getProperty().

Make ubic/gemma/version.properties mandatory. It should always be there
since it's generated in the build. Move its generation in the
generate-resources so that it is created in the compile lifecycle,
before test is run.

Update ConfigurationTest to also verify that the environment is
populated with the settings and only use classpath-supplied properties.