Relocate some of the statement logic back into FactorValueService
Add a createStatement() which allows us to apply ACLs at the FactorValue
and check if an identical statement already exist.

Now that we have a StatementDao, we can use it instead of having to put
database logic in the FactorValueDao.

Move cloning back into SplitExperimentServiceImpl since there's nothing
complicated about it anymore.