Improve migration of old-style characteristics
Relocate migration logic in a separate service so that we can migrate
each factor value in individual transactions and test it more easily.

Migrate unmentioned characteristics in a factor value as subject-only
statements.

Migrate all remaining factor values with a single characteristic as
subject-only statements. All other unmigrated FVs produce a warning.

Add a flag for migrated old-style characteristics and set it when they
are being used in a statement.

Add missing fixture for testing FV migration.