Keep track of FVs that need attention
Add a NEEDS_ATTENTION column in FACTOR_VALUE  and a FactorValueNeedsAttentionEvent
to keep track of individual factor values that need attention.

Add a flag in the migrateFactorValues CLI to migrate non-trivial cases
which are then flagged as need attention.

Add a mechanism in AbstractCLI for prompting for user confirmation and
aborting the current work.

Expose in the curator UI which FVs need attention.