Eliminate expressionExperiment access from SingleCellExpressionDataVector
Use proper ownership semantic for single-cell data vectors as to not
repeat the mistake for raw and process vectors.

Relocate the EE field in BulkExpressionDataVector.

Revert "Eliminate expressionExperiment access from SingleCellExpressionDataVector"

This reverts commit adaf534770d368b376067ec867f5530bc0f73b6c.

Turns out having this setup that way produces additional update
statements to persist the relation, so it's not worth it.