Add support for multiple cell type labelling
Makes it so that a single-cell dimension can have multiple cell type
labelling. The cell type labelling is represented as a subclass of an
Analysis and its protocol describe the labelling process in details.

Create an experimental factor for the cell type when labels are assigned
to the preferred single-cell vectors. The factor is re-created when new
labels are assigned and removed when those labels are removed.