Select a representative characteristic for each experiment and level
The representative characteristic of each group is always the one with
the minimal ID. This ensures that adding a characteristic to an
experiment that already exists cannot result in a new row getting
inserted in the EE2C table.

Remove LEVEL from the primary key, it's already guaranteed that
characteristic IDs are unique per level.