Ensure that all our executors are being shutdown
This has been a long-standing issue that prevents the JVM from shutting
down.

Executors are not using daemon threads by default, so we can either use
a thread factory that does that or consistently shut it down. The latter
approach is taken.

In some cases, use shutdownNow() to accelerate shutting down when
non-critical components and disposed.