Few more rounds of optimization for ontology caches
Use a read-only lock when accessing the ontology cache initially and
then a read-write lock to compute.

Eliminate the k-1,k-2 subset strategy, it's completely inefficient in
practice and the caches are so small that it's more performance to just
enumerate the keys.

Make parents/children cache eternal. These relationships are immutable
once the ontology is loaded.