No need to use a TreeMap to aggregate categories and annotations counts
Those entities are non-persistent, so regular coalescing rules of URIs
and labels apply. Also, because this is collected in a mapping, order is
not important.

Also fix bugs where a term with different labels appear twice in the
output because of the sorting by label.

Exclude the id from Characteristic's hash code