Make sure that DNF terms containing spaces or special characters are properly escaped
Full-text search against the ontology is done only using exact matches
on terms and limited boolean logic. Thus, when a subclause contains
special characters or spaces, those must be escaped or quoted.

Rename extractDnf to extractTermsDnf to make it clear this this will
only pull down the terms and not prefix, wildcards, etc.