Improve output of AnnotationWebService
Honor the same query parsing used for searching datasets when searching
annotations. This include not stripping quotes when performing the
search against ontologies.

Use LuceneQueryUtils for parsing wildcard and URI queries.

Relocate escapeLike in QueryUtils and reuse it consistently.

Use a QueriedAndFilteredAndPaginated type of payload such that the query
can be presented in the output instead of the matching dataset IDs.

Finish early if one of the query return no result as the intersection
will be empty.

Describe how multiple query parameters are handled in
AnnotationWebService. Fix #1100.

Fix duplicated annotation results. Those are now collected in a
LinkedHashSet. Fix #1084.