Filter system properties that are used for placeholders
Only allow properties declared in either default.properties or
project.properties to be specified.

System properties should be prefixed with 'gemma.' to prevent
conflicting options. A warning is produced if a declared property is not
prefixed. Those will be ignored in the future.

Make PropertySources a singleton bean so that it is only loaded once on
startup.

Share some logic between the deprecated Settings and the new
SettingsConfig.