handle PROBLEMATIC_BATCH_INFO_FAILURE in display (treat as "no batch info")
Note: This flag is somewhat different from UNINFORMATIVE_HEADERS_FAILURE - the headers were present, but we failed to get batch information from them, but more likely to be due to a problem with the FASTQ header file missing rows or something like that, vs UNINFORMATIVE_HEADERS_FAILURE is used when the only problem is the headers don't contain lane etc. info.