cli: Refresh dataset after batch information is submitted
Add a minimalistic client for the REST API.

Add a CLI for refreshing experiments.

Add support for loading EE by reference so that -all can run very
quickly if all that is needed are IDs.

Few improvements when specifying EEs by ID. This is now supported in
both CLI argument and files.