Only scan ubic.gemma.core and ubic.gemma.persistence for components
This prevent scanning over-and-over again the same classes on startup
and make it possible to exclude the regex rule for Gemma Web.

Add a component-scan directive for Gemma REST which is no longer
covered.

Remove @Component annotation to CLIs to prevent them from being picked
up by gemma-core component scanner.

Drop default filters for gemma-cli component scanner, just look for CLIs
that are not TestComponent.