Align results by gene with the development branch
Grouping by result set ID, handling of non-specific probes, and a few
other changes weren't included in the initial backport.

Rename the endpoint and path to match what is in the development trunk.

Add offset/limit parameters based on the dataset IDs instead of the
results being retrieved. This means we are sorting the payload by
'sourceExperimentId'.

Expose probe IDs which do not cost anything.