WIP
Introduce a hierarchy of BioMaterial such that a sample can be derived
from another sample and inherit its characteristics, treatments and
factor values.

Prevent splitting an experiment that uses a hierarchy of biomaterial for
now. This might be implemented in the future.

Add a separate field in QuantitationType to identify preferred
single-cell QT.