Migrate to Hibernate 4.0 XML schema
Patch LocalSessionFactoryBean to use an entity resolver that can load
Hibernate XSD schemas from the classpath.

Due to the new schemas, we cannot use Hibernate hbm2ddm ant tasks
because they cannot use the XSD entity resolver. Instead, the database
initialization is reimplemented as a Spring DataSourceInitializer and a
bunch of DataSourcePopulator.

Remove all the build logic for generating DDL SQL files, those are now
done at runtime in tests.

Add a CLI for populating the test database manually. This CLI is only
available when the testdb profile is active to prevent accidental
usages.

Fix the PersistentDummyObjectHelper which incorrectly retains entities
in static properties. This is breaking the test context when the
database is regenerated.