Multiple improvements for GeoBrowser and GeoGrabberCli
When rewinding, start exactly on the dataset that meet the criteria
instead of its chunk.

Pass the NCBI API key by constructor so that we can use @Value to inject
it in the CLIs.

Utilize [ORGN] to filter by taxa on GEO side instead of wasting time on
the client-side. This is useful to reduce the amount of records we need
to seek.

Properly quote terms in Entrez search queries.

Checkout the SimpleRetry implementation from the feature-single-cell
branch that is used for retrying FTP downloads.

Add chunkSize parameter

Filter series types in the Entrez query.