Initial implementation of MexSingleCellMatrixWriter
Add support for writing MEX data as a TAR archive or to a directory. The
latter is much more efficient because we can stream the vectors.

Relocate TSV-related utilities under a new package.

Add methods for retrieving platforms used by a specific set of vectors

This allow one to populate gene mappings for a set of vectors regardless
of the BioAssay's platforms.