More improvements for EE-manipulating CLIs
Only call processBioAssaySets() if more than one experiment is matched.
This way, subclasses can assume that if this is called, options that are
designed for a single-experiment should not be used.

Add preprocessBioAssaySets() to given an opportunity to perform
some work and filtering in a subclass.

Allow subclasses to override doAuthenticatedWork(), some CLIs have
switches that makes them completely ignore the experiments.

Undo the flawed logic for detecting if a success/error object was added,
this does not work if processBioAssay() uses the batch task executor.