rest: Few more improvements for locked files
Make all APIs in ExpressionDataFileService return LockedPath and adjust
the usage to make sure that the file is read while the lock is held.

Add a PathProvider to use Java 8 Path API. This will be removed once we
update Jersey to a more recent version.

Add a LockedPathProvider to write the content of a LockedPath using the
PathProvider and release the lock afterward.

Add a MexBundlerProvider to automatically bundle MEX-structured folders
as TAR archive when the media type is set.