Few improvements for the static assets server
Add resolveUrl and use that in the tags to generate link to resources.

Add a list of allowed directories to serves assets from. It does not
really do much security-wise, but it will catch unintended usages of the
JSP tags.

Add a useful message when the development profile is active and the
server is not running.