Fix visualization diff. ex results by subset
Add a new interface BioAssaySetValueObject and use it in the code so
that we don't assume that an ExpressionExperimentSubsetValueObject
derives from an ExpressionExperimentValueObject.

Reintroduce the DoubleVectorValueObject cache. Since we're not caching
processed vector in Hibernate anymore, this now make sense to do.

Make sure that operations on ProcessedExpressionDataVectorService evict
the vectors and vectors by gene caches.

Make sure that we create copies when setting a minimum P-value or rank
on a vector as to not contaminate the content of the cache. Ideally,
those models should be immutable.

Regenerate the DWR client code as we're adding a new model.