Fix spring.profiles.active JVM setting not being honored
It looks like spring.profiles.active servlet init parameter got fixed,
but I'm not exactly sure how since we haven't updated Spring. In any
case, the value is overriding the JVM setting, so we cannot use it in
web.xml.

Instead, simply add the web profile when initializing the context.