Enable sendfile for the rest API
Check if org.apache.tomcat.sendfile.support is set and produce a warning
if tomcat.sendfile.enabled is set as well.

Drop usage of LockedPath, this was causing dangling locks that were
never released. Moreover, it's not compatible with using sendfile as we
delegate the download to something we do not control. Fix #1352.

Improve the test for downloading raw data files which exercise this
feature.