Allow redo/extend analysis to be supplied a base configuration
This makes the -redo flag honor other options passed from the CLI.

No need to set weights since it is already set in the base
configuration.

Cleanup how we read configuration options for DEA from the CLI.

Add an --output-dir option to diffExAnalyze to generate the result file
in a specified directory.

Add a CLI for writing DEA results as they appear in the database.