Add an option to ignore the cell type factor when loading AnnData
The AnnData loader will always try to check if a cell type factor was
set when loading cell type assignments. This check is quite useful when
the cell type cannot be detected or was mistyped.

This adds an option so that we can bypass this check and load AnnData
without CTAs.